pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-srs-assistant"
version = "0.1.0"
description = "Agentic Software Requirements Specification (SRS) assistant using AI agents"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Agentic SRS Assistant Team"}
]
license = "MIT"
keywords = ["srs", "requirements", "ai", "agentic", "langgraph", "streamlit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core dependencies
    "streamlit>=1.28.0",
    "langgraph>=0.0.20",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "langchain>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    # Optional dependencies (can be installed separately)
    "langfuse>=2.0.0; extra == 'observability'",
    "psycopg2>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]
observability = [
    "langfuse>=2.0.0",
]
all = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
    "langfuse>=2.0.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langchain.*",
    "streamlit.*",
    "psycopg2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
